requirements.txt
alembic==1.4.3
ansible==2.10.1
ansible-runner==1.4.6
Authlib==0.14.3
httpx==0.15.5
fastapi==0.92.0
fastapi-pagination[sqlalchemy]==0.11.4
ipython==7.18.1
itsdangerous==1.1.0
redis==3.5.3
Jinja2==3.0.3
psycopg2-binary==2.9.3
pytest==6.1.0
requests==2.24.0
SQLAlchemy==1.4.46
sqlalchemy-utils==0.38.2
uvicorn[standard]
passlib==1.7.2
bcrypt==3.2.0
python-multipart==0.0.5
pyjwt==1.7.1
psutil==5.7.2
sentry-sdk==1.5.5
dnspython==2.1.0
huey==2.4.2
fabric==2.7.1
patchwork==1.0.1
strawberry-graphql[fastapi]
asyncpg
sse-starlette==1.3.3
websockets==10.4
aiofiles==23.1.0